+++ extra_foam/special_suite/xas_tim_w.py
@@ -348,7 +348,7 @@ class XasTimWindow(_SpecialAnalysisBase):
 
         cw = self.centralWidget()
         cw.addWidget(right_panel)
-        cw.setSizes([self._TOTAL_W / 4, 3 * self._TOTAL_W / 4])
+        cw.setSizes([int(self._TOTAL_W / 4), int(3 * self._TOTAL_W / 4)])
 
         self.resize(self._TOTAL_W, self._TOTAL_H)
